+++ src/main/java/com/safetynet/alerts/web/controller/MedicalRecordController.java
@@ -1,7 +1,8 @@
 package com.safetynet.alerts.web.controller;
 
-
 import org.tinylog.Logger;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,8 +26,8 @@ import com.safetynet.alerts.web.service.MedicalRecordService;
 @RequestMapping("/medicalRecord")
 public class MedicalRecordController {
 
-
     private final MedicalRecordService medicalRecordService;
+    private BeanService beanService = new BeanService();
 
     public MedicalRecordController(MedicalRecordService medicalRecordService) {
         this.medicalRecordService = medicalRecordService;
@@ -40,9 +41,18 @@ public class MedicalRecordController {
      *                      record to be added.
      */
     @PostMapping("")
-    public void addMedicalRecord(@RequestBody MedicalRecord medicalrecord) {
-        medicalRecordService.saveMedicalRecord(medicalrecord);
-        Logger.info("Request : " + BeanService.getCurrentMethodName() + "().");
+    public ResponseEntity<String> addMedicalRecord(@RequestBody MedicalRecord medicalrecord) {
+        Logger.info("Request " + BeanService.getCurrentMethodName() + ".");
+        Boolean areFieldsAreNull = beanService.areFieldsNullExceptId(medicalrecord);
+        if (areFieldsAreNull) {
+            Logger.error("Answer " + BeanService.getCurrentMethodName() + " : content is incorrect.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Medicalrecord content is incorrect.");
+        } else {
+            medicalRecordService.saveMedicalRecord(medicalrecord);
+            Logger.info("Answer " + BeanService.getCurrentMethodName() + " : medicalrecord added successfully.");
+            return ResponseEntity.status(HttpStatus.CREATED).body("Medicalrecord added successfully.");
+        }
+
     }
 
     /**
@@ -57,27 +67,37 @@ public class MedicalRecordController {
      *                         information.
      */
     @PutMapping("/{firstName}/{lastName}")
-    public void updateMedicalRecord(@PathVariable("firstName") String firstName,
+    public ResponseEntity<String> updateMedicalRecord(@PathVariable("firstName") String firstName,
             @PathVariable("lastName") String lastName, @RequestBody MedicalRecord newMedicalRecord) {
-        MedicalRecord oldMedicalrecord = medicalRecordService.findMedicalRecordByFirstNameAndLastName(firstName,
-                lastName);
-        if (oldMedicalrecord != null) {
-            try {
-                MedicalRecord updatMedicalrecord = BeanService
-                        .updateBeanWithNotNullPropertiesFromNewObject(oldMedicalrecord, newMedicalRecord);
-                updatMedicalrecord.setId(oldMedicalrecord.getId());
-                updatMedicalrecord.setFirstName(firstName);
-                updatMedicalrecord.setLastName(lastName);
-                medicalRecordService.saveMedicalRecord(updatMedicalrecord);
-                Logger.info("Request : " + BeanService.getCurrentMethodName() + " with firstName and lastName : " + firstName
-                        + " " + lastName + ".");
-            } catch (Exception e) {
-                Logger.error("Request : " + BeanService.getCurrentMethodName() + " with firstName and lastName : " + firstName
-                        + " " + lastName + ", genereted this exception : " + e);
-            }
+        Logger.info("Request " + BeanService.getCurrentMethodName() + " : with this first and last name = " + firstName
+                + " "
+                + lastName + ".");
+        Boolean areFieldsAreNull = beanService.areFieldsNullExceptId(newMedicalRecord);
+        if (areFieldsAreNull) {
+            Logger.error("Answer " + BeanService.getCurrentMethodName() + " : content is incorrect.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("MedicalRecord content is incorrect.");
         } else {
-            Logger.error("Request : " + BeanService.getCurrentMethodName() + " with firstName and lastName : " + firstName
-                    + " " + lastName + ".");
+            MedicalRecord oldMedicalrecord = medicalRecordService.findMedicalRecordByFirstNameAndLastName(firstName,
+                    lastName);
+            if (oldMedicalrecord != null) {
+                try {
+                    MedicalRecord updatMedicalrecord = BeanService
+                            .updateBeanWithNotNullPropertiesFromNewObject(oldMedicalrecord, newMedicalRecord);
+                    updatMedicalrecord.setId(oldMedicalrecord.getId());
+                    updatMedicalrecord.setFirstName(firstName);
+                    updatMedicalrecord.setLastName(lastName);
+                    medicalRecordService.saveMedicalRecord(updatMedicalrecord);
+                    Logger.info(
+                            "Answer " + BeanService.getCurrentMethodName() + " : modified successfully.");
+                    return ResponseEntity.status(HttpStatus.OK).body("MedicalRecord modified successfully.");
+                } catch (Exception e) {
+                    Logger.error("Answer " + BeanService.getCurrentMethodName() + " : threw an exception.");
+                    return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Request threw an exception.");
+                }
+            } else {
+                Logger.error("Answer " + BeanService.getCurrentMethodName() + " : first and last name doesn't match.");
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("First and last name doesn't match.");
+            }
         }
     }
 
@@ -92,10 +112,20 @@ public class MedicalRecordController {
      */
     @Transactional
     @DeleteMapping("/{firstName}/{lastName}")
-    public void deleteMedicalRecord(@PathVariable("firstName") String firstName,
+    public ResponseEntity<String> deleteMedicalRecord(@PathVariable("firstName") String firstName,
             @PathVariable("lastName") String lastName) {
-        medicalRecordService.deleteMedicalRecordByFirstNameAndLastName(firstName, lastName);
-        Logger.info("Request : " + BeanService.getCurrentMethodName() + " with firstName and lastName : " + firstName
-                + " " + lastName + ".");
+        Logger.info(
+                "Request " + BeanService.getCurrentMethodName() + " : with this first and last name  = " + firstName
+                        + " " + lastName + ".");
+        MedicalRecord oldMedicalrecord = medicalRecordService.findMedicalRecordByFirstNameAndLastName(firstName,
+                lastName);
+        if (oldMedicalrecord != null) {
+            medicalRecordService.deleteMedicalRecordByFirstNameAndLastName(firstName, lastName);
+            Logger.info("Answer " + BeanService.getCurrentMethodName() + " : deleted successfully.");
+            return ResponseEntity.status(HttpStatus.OK).body("MedicalRecord deleted successfully.");
+        } else {
+            Logger.error("Answer " + BeanService.getCurrentMethodName() + " : doesn't match with any medicalRecord.");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("MedicalRecord address doesn't match.");
+        }
     }
 }
